requirements.txt
argparse==1.2.1
paho-mqtt==1.1
pyserial==3.0
wsgiref==0.1.2
